+++ packages/@pollyjs/adapter-puppeteer/CHANGELOG.md
@@ -3,6 +3,44 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.1.4"></a>
+## [1.1.4](https://github.com/netflix/pollyjs/tree/master/packages/@pollyjs/adapter-puppeteer/compare/@pollyjs/adapter-puppeteer@1.1.3...@pollyjs/adapter-puppeteer@1.1.4) (2018-08-22)
+
+
+### Bug Fixes
+
+* Puppeteer 1.7.0 support ([#100](https://github.com/netflix/pollyjs/tree/master/packages/[@pollyjs](https://github.com/pollyjs)/adapter-puppeteer/issues/100)) ([e208b38](https://github.com/netflix/pollyjs/tree/master/packages/@pollyjs/adapter-puppeteer/commit/e208b38))
+
+
+
+
+<a name="1.1.3"></a>
+## [1.1.3](https://github.com/netflix/pollyjs/tree/master/packages/@pollyjs/adapter-puppeteer/compare/@pollyjs/adapter-puppeteer@1.1.2...@pollyjs/adapter-puppeteer@1.1.3) (2018-08-12)
+
+
+
+
+**Note:** Version bump only for package @pollyjs/adapter-puppeteer
+
+<a name="1.1.2"></a>
+## [1.1.2](https://github.com/netflix/pollyjs/tree/master/packages/@pollyjs/adapter-puppeteer/compare/@pollyjs/adapter-puppeteer@1.1.1...@pollyjs/adapter-puppeteer@1.1.2) (2018-08-12)
+
+
+### Bug Fixes
+
+* **adapter-puppeteer:** Do not intercept CORS preflight requests ([#90](https://github.com/netflix/pollyjs/tree/master/packages/[@pollyjs](https://github.com/pollyjs)/adapter-puppeteer/issues/90)) ([53ad433](https://github.com/netflix/pollyjs/tree/master/packages/@pollyjs/adapter-puppeteer/commit/53ad433))
+
+
+
+
+<a name="1.1.1"></a>
+## [1.1.1](https://github.com/netflix/pollyjs/tree/master/packages/@pollyjs/adapter-puppeteer/compare/@pollyjs/adapter-puppeteer@1.1.0...@pollyjs/adapter-puppeteer@1.1.1) (2018-08-09)
+
+
+
+
+**Note:** Version bump only for package @pollyjs/adapter-puppeteer
+
 <a name="1.1.0"></a>
 # [1.1.0](https://github.com/netflix/pollyjs/tree/master/packages/@pollyjs/adapter-puppeteer/compare/@pollyjs/adapter-puppeteer@1.0.0...@pollyjs/adapter-puppeteer@1.1.0) (2018-07-26)