+++ packages/@pollyjs/persister-rest/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.0.5"></a>
+## [1.0.5](https://github.com/netflix/pollyjs/tree/master/packages/@pollyjs/persister-rest/compare/@pollyjs/persister-rest@1.0.4...@pollyjs/persister-rest@1.0.5) (2018-08-22)
+
+
+
+
+**Note:** Version bump only for package @pollyjs/persister-rest
+
+<a name="1.0.4"></a>
+## [1.0.4](https://github.com/netflix/pollyjs/tree/master/packages/@pollyjs/persister-rest/compare/@pollyjs/persister-rest@1.0.3...@pollyjs/persister-rest@1.0.4) (2018-08-12)
+
+
+
+
+**Note:** Version bump only for package @pollyjs/persister-rest
+
+<a name="1.0.3"></a>
+## [1.0.3](https://github.com/netflix/pollyjs/tree/master/packages/@pollyjs/persister-rest/compare/@pollyjs/persister-rest@1.0.2...@pollyjs/persister-rest@1.0.3) (2018-08-12)
+
+
+
+
+**Note:** Version bump only for package @pollyjs/persister-rest
+
+<a name="1.0.2"></a>
+## [1.0.2](https://github.com/netflix/pollyjs/tree/master/packages/@pollyjs/persister-rest/compare/@pollyjs/persister-rest@1.0.1...@pollyjs/persister-rest@1.0.2) (2018-08-09)
+
+
+
+
+**Note:** Version bump only for package @pollyjs/persister-rest
+
 <a name="1.0.1"></a>
 ## [1.0.1](https://github.com/netflix/pollyjs/tree/master/packages/@pollyjs/persister-rest/compare/@pollyjs/persister-rest@1.0.0...@pollyjs/persister-rest@1.0.1) (2018-07-26)